Add files via upload
This project takes a look at Airbnb prices found throughout the Seattle area in 2016. The visual created with this data provides the average price of an Airbnb rental in each zip code, the total revenue per week throughout the entire year, and the average nightly price per number of bedrooms, along with how many listings have each number of bedrooms. Although the data is outdated if completed with new data this could help Airbnb hosts rent their property for a value which is in line with other hosts' prices and also could help hosts find which houses and features could provide them with the most revenue.